requirements.txt
psycopg2-binary>=2.9
requests>=2.28
pytest>=7.0
streamlit>=1.30
pandas>=2.0
plotly>=5.18
